Authorization implemented in all sales APIs

+++ src/routes/salesRoutes.js
@@ -1,24 +1,31 @@
 import express from "express";
 import { v4 as uuidv4 } from 'uuid';
 import { salesRegister } from "../models/SalesModel.js";
+import { authenticate } from "../middleware/middleware.js";
+import { getUserIdFromToken } from "../helper/generateToken.js";
 
 const router = express.Router();
 
-router.post("/api/usersales/addSalesRegister", async (req, res) => {
+router.post("/api/usersales/addSalesRegister", authenticate, async (req, res) => {
     try {
+        const token = req.header('Authorization');
+        const userId = getUserIdFromToken(token)
         const salesRegisterData = req.body;
         const findSaleRegId = uuidv4();
         salesRegisterData["saleRegId"] = findSaleRegId;
+        salesRegisterData["userId"] = userId;
         const createSalesData = await salesRegister.create(salesRegisterData);
         return res.status(200).json({ success: true, message: "Sales Register Data Added Successfully!", saleRegId: createSalesData.saleRegId });
     } catch (error) {
-        return res.status(500).json({ success: false, message: "Internal Server Error!" });
+        return res.status(500).json({ success: false, message: error.message });
     }
 });
 
-router.post("/api/usersales/getAllSalesList", async (req, res) => {
+router.post("/api/usersales/getAllSalesList", authenticate, async (req, res) => {
     try {
-        const findAllSalesList = await salesRegister.find({}); 
+        const token = req.header('Authorization');
+        const userId = getUserIdFromToken(token);
+        const findAllSalesList = await salesRegister.find({ userId: userId }, { _id:0, __v:0 }); 
         if (findAllSalesList) {
             return res.status(200).json({ success: true, message: findAllSalesList });
         } else {
@@ -29,11 +36,13 @@ router.post("/api/usersales/getAllSalesList", async (req, res) => {
     }
 });
 
-router.post("/api/usersales/getSalesRegister", async (req, res) => {
+router.post("/api/usersales/getSalesRegister", authenticate, async (req, res) => {
     try {
+        const token = req.header('Authorization');
+        const userId = getUserIdFromToken(token);
         const salesRegisterData = req.body;
         const getUserSalesId = salesRegisterData.saleRegId;
-        const findSalesUserId = await salesRegister.findOne({ saleRegId: getUserSalesId });
+        const findSalesUserId = await salesRegister.findOne({ userId: userId, saleRegId: getUserSalesId }, { __v:0, _id:0 });
         if (findSalesUserId) {
             return res.status(200).json({ success: true, message: findSalesUserId });
         } else {
@@ -44,11 +53,13 @@ router.post("/api/usersales/getSalesRegister", async (req, res) => {
     }
 });
 
-router.put("/api/usersales/updateSalesRegData", async (req, res) => {
+router.put("/api/usersales/updateSalesRegData", authenticate, async (req, res) => {
     try {
+        const token = req.header('Authorization');
+        const userId = getUserIdFromToken(token);
         const salesRegisterData = req.body;
         const getUserSalesId = salesRegisterData.saleRegId;
-        const findSalesUserId = await salesRegister.findOneAndUpdate({ saleRegId: getUserSalesId }, { $set: salesRegisterData });
+        const findSalesUserId = await salesRegister.findOneAndUpdate({ userId: userId, saleRegId: getUserSalesId }, { $set: salesRegisterData });
         if (findSalesUserId) {
             return res.status(200).json({ success: false, message: "Sales Register Data Id Is Updated Successfully!" });
         } else {
@@ -59,11 +70,13 @@ router.put("/api/usersales/updateSalesRegData", async (req, res) => {
     }
 });
 
-router.delete("/api/usersales/deleteSalesData", async (req, res) => {
+router.delete("/api/usersales/deleteSalesData", authenticate, async (req, res) => {
     try {
+        const token = req.header('Authorization');
+        const userId = getUserIdFromToken(token); 
         const salesRegisterData = req.body;
         const getUserSalesId = salesRegisterData.saleRegId;
-        const findSalesUserId = await salesRegister.findOneAndDelete({ saleRegId: getUserSalesId });
+        const findSalesUserId = await salesRegister.findOneAndDelete({ userId: userId, saleRegId: getUserSalesId });
         if (findSalesUserId) {
             return res.status(200).json({ success: false, message: "Sales Register Data Id Is Deleted Successfully!" });
         } else {
